style(@layer2/core, @layer2/wallet): apply format
affects: @layer2/core, @layer2/wallet
