+++ packages/core/src/block.ts
@@ -8,7 +8,8 @@ import BigNumber = utils.BigNumber
 import { TOTAL_AMOUNT } from './helpers/constants'
 import { Hash, Address } from './helpers/types'
 import { MapUtil } from './utils/MapUtil'
-import { SegmentedBlock, ExclusionProof } from './models/SegmentedBlock'
+import { SegmentedBlock } from './models/SegmentedBlock'
+import { ExclusionProof } from './models/ExclusionProof'
 import { StateUpdate } from './StateUpdate'
 import { PredicatesManager } from './predicates'
 import { SegmentNode } from './models'
