+++ packages/core/src/models/index.ts
@@ -1,1 +1,4 @@
+export { ExclusionProof } from './ExclusionProof'
 export { SegmentNode } from './SegmentNode'
+export { SegmentedBlock } from './SegmentedBlock'
+export * from './swap'
