+++ packages/core/src/predicates/PredicatesManager.ts
