+++ packages/core/src/utils/ChamberResult.ts
@@ -1,0 +1,8 @@
+import { ChamberError } from './error'
+
+export interface ChamberResult<T> {
+  isOk(): boolean
+  isError(): boolean
+  ok(): T
+  error(): ChamberError
+}
