+++ packages/wallet/src/client/PlasmaClient.ts
@@ -10,16 +10,7 @@ import {
   SwapRequest,
   SignedTransaction
 } from '@layer2/core'
-
-export class FastTransferResponse {
-  public sig: string
-  public tx: SignedTransaction
-
-  constructor(sig: string, tx: SignedTransaction) {
-    this.sig = sig
-    this.tx = tx
-  }
-}
+import { FastTransferResponse } from '../models/FastTransferResponse'
 
 export class PlasmaClient {
   public static deserialize<T>(
