+++ packages/wallet/src/history/index.ts
@@ -1,0 +1,3 @@
+export { PlasmaBlockHeader } from './PlasmaBlockHeader'
+export { SegmentHistory } from './SegmentHistory'
+export { SegmentHistoryManager } from './SegmentHistoryManager'
