+++ packages/wallet/src/models/index.ts
@@ -2,3 +2,4 @@ export { Exit } from './exit'
 export { WaitingBlockWrapper } from './WaitingBlockWrapper'
 export { TokenType } from './TokenType'
 export { UserAction, UserActionUtil } from './UserAction'
+export { FastTransferResponse } from './FastTransferResponse'
