+++ packages/wallet/test/mock/MockNetworkClient.ts
@@ -1,6 +1,4 @@
-import {
-  INetworkClient,
-} from '../../src/client'
+import { INetworkClient } from '../../src/client'
 
 export class MockNetworkClient implements INetworkClient {
   public request(methodName: string, args: any) {
