+++ packages/wallet/test/mock/MockPubsubClient.ts
@@ -1,12 +1,13 @@
-import {
-  IPubsubClient,
-  SubscribeHandler
-} from '../../src/client'
+import { IPubsubClient, SubscribeHandler } from '../../src/client'
 
 export class MockPubsubClient implements IPubsubClient {
   public publish(topic: string, message: string) {
     return true
   }
-  public subscribe(topic: string, event: SubscribeHandler): void {}
-  public unsubscribe(topic: string, handler: SubscribeHandler): void {}
+  public subscribe(topic: string, event: SubscribeHandler): void {
+    return
+  }
+  public unsubscribe(topic: string, handler: SubscribeHandler): void {
+    return
+  }
 }
